+++ spec/app/domain/secrets/cache/redis_handler_spec.rb
@@ -69,6 +69,29 @@ describe Secrets::RedisHandler do
     end
   end
 
+  context "Secret update - error handling" do
+    let(:key) {"data/key"}
+
+    before do
+      Rails.cache.clear
+    end
+
+    it "Secret is deleted from Redis if update failed" do
+      controller.create_redis_secret(key, "1", "")
+      allow(controller).to receive(:create_redis_secret).and_return('false')
+      expect(controller).to receive(:delete_redis_secret).with(key, suppress_error: false)
+      controller.update_redis_secret(key, "2")
+    end
+
+    it "Error is raised if deletion fails during update" do
+      controller.create_redis_secret(key, "1", "")
+      allow(controller).to receive(:create_redis_secret).and_return('false')
+      allow(controller).to receive(:delete_redis_secret).with(key, suppress_error: false).and_raise(RuntimeError)
+      expect{ controller.update_redis_secret(key, "2") }.to raise_error
+    end
+
+  end
+
   class Controller
     include Secrets::RedisHandler
   end